+++ indico/modules/events/registration/client/js/form/FormItem.jsx
@@ -107,6 +107,22 @@ export default function FormItem({
 
   const showAsRequired = meta.alwaysRequired || isRequired;
   const inputRequired = !isManagement && showAsRequired;
+  const htmlId = `input-${inputProps.id}`;
+
+  const fieldControls =
+    InputComponent && !meta.customFormItem ? (
+      <>
+        {retentionPeriodIcon}
+        <InputComponent
+          isRequired={inputRequired}
+          disabled={disabled}
+          isPurged={showPurged}
+          htmlId={htmlId}
+          {...inputProps}
+        />
+      </>
+    ) : null;
+
   return (
     <div
       styleName={`form-item ${toClasses({
@@ -126,18 +142,29 @@ export default function FormItem({
               disabled={disabled}
               isPurged={showPurged}
               retentionPeriodIcon={retentionPeriodIcon}
+              htmlId={htmlId}
               {...inputProps}
             />
           ) : (
             <Form.Field required={showAsRequired} styleName="field">
-              <label style={{opacity: disabled ? 0.8 : 1, display: 'inline-block'}}>{title}</label>
-              {retentionPeriodIcon}
-              <InputComponent
-                isRequired={inputRequired}
-                disabled={disabled}
-                isPurged={showPurged}
-                {...inputProps}
-              />
+              {meta.renderAsFieldset ? (
+                <fieldset>
+                  <legend style={{opacity: disabled ? 0.8 : 1, display: 'inline-block'}}>
+                    {title}
+                  </legend>
+                  {fieldControls}
+                </fieldset>
+              ) : (
+                <>
+                  <label
+                    htmlFor={htmlId}
+                    style={{opacity: disabled ? 0.8 : 1, display: 'inline-block'}}
+                  >
+                    {title}
+                  </label>
+                  {fieldControls}
+                </>
+              )}
             </Form.Field>
           )
         ) : (