+++ indico/modules/events/registration/client/js/form/fields/CheckboxInput.jsx
@@ -23,6 +23,7 @@ import styles from '../../../styles/regform.module.scss';
 
 export default function CheckboxInput({
   id,
+  htmlId,
   htmlName,
   disabled,
   title,
@@ -35,9 +36,11 @@ export default function CheckboxInput({
 }) {
   const currency = useSelector(getCurrency);
   const existingValue = useSelector(state => getFieldValue(state, id));
+  const checkboxId = `${htmlId}-checkbox`;
 
   return (
     <FinalCheckbox
+      id={checkboxId}
       fieldProps={{className: `${styles.field} ${showAsRequired ? 'required' : ''}`}}
       name={htmlName}
       label={title}
@@ -65,6 +68,7 @@ export default function CheckboxInput({
 
 CheckboxInput.propTypes = {
   id: PropTypes.number.isRequired,
+  htmlId: PropTypes.string.isRequired,
   htmlName: PropTypes.string.isRequired,
   disabled: PropTypes.bool,
   title: PropTypes.string.isRequired,